+++ hms_mastodon/strings.py
@@ -1,6 +1,11 @@
-TWAUM_OPEN = "INFO : notre espace est tout ouvert, " \
-             "n’hésitez pas à passer si vous le voulez/pouvez ! https://haum.org"
+import emoji
 
-TWAUM_CLOSED = "Fin de session ! Jetez un œil à notre agenda sur " \
+TWAUM_OPEN_SHORT = emoji.emojize(":blue_circle: L'espace est ouvert")
+TWAUM_CLOSED_SHORT = emoji.emojize(":red_circle: Fin de session")
+
+TWAUM_OPEN = "N’hésitez pas à passer si vous le voulez/pouvez ! " \
+             "https://haum.org"
+
+TWAUM_CLOSED = "Jetez un œil à notre agenda sur " \
                "https://haum.org pour connaître les prochaines ou " \
                "surveillez notre fil Mastodon."