+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`__.
 [Unreleased]
 ------------
 
+- Emojis and shorter toot messages (with spoiler feature) in order to not spam
+  the toot timeline of users
+
 [1.1] - 2017-04-16
 ------------------