+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`__.
 [Unreleased]
 ------------
 
+[1.2] - 2017-05-08
+------------------
+
 - Streaming API support for mentions
 - Emojis and shorter toot messages (with spoiler feature) in order to not spam
   the toot timeline of users